URGENT: Fix live production outage of /api/radar/scores -- entire Radar Screen was returning 500 errors
User reported the Radar Screen wasn't populating at all. Real
production logs showed GET /api/radar/scores returning 500, with the
actual error: ValueError: [TypeError("'numpy.int64' object is not
iterable"), TypeError('vars() argument must have __dict__
attribute')] -- a JSON serialization failure, breaking the entire
endpoint, not just the new Methodology Breakdown feature.

Traced precisely rather than guessed: tested each of the three new
verdict engines' real output directly against json.dumps(), isolating
the exact failure to LivermoreVerdictEngine specifically.
failure_frequency_score() computes failures = (...).sum() on a
pandas boolean Series, which returns numpy.int64, not a plain Python
int -- this well-known pandas/numpy gotcha then flowed unconverted
through min() and the weighted score calculation into the final
dataclass, which FastAPI's jsonable_encoder cannot serialize.

Fixed at the source: cast to int() explicitly in
failure_frequency_score(). Also added a second, defensive layer:
_sanitize_numpy(), a recursive sanitizer applied to all three
verdicts before they're attached to any row, converting any numpy
scalar type to its plain Python equivalent -- protects against this
exact class of bug even if these engines (or others attached here
later) are modified again and reintroduce a similar leak.

Verified precisely: re-ran the same direct json.dumps() test against
all three engines' real output, confirming all three now pass. Then
ran the complete, real (unmocked engines) get_radar_scores() flow
end-to-end and confirmed the full response is genuinely
JSON-serializable -- the exact failure mode that broke production is
now directly, verifiably fixed, not just assumed fixed. Full test
suite (70 tests) passes.

This needs sigmalytic-backend redeployed immediately given the
severity -- the entire Radar Screen is currently down in production.

+++ backend/research_engine/livermore_verdict_engine.py
@@ -124,7 +124,15 @@ class LivermoreVerdictEngine:
             (recent["close"] > support)
         ).sum()
 
-        return min(100, failures * 20)
+        # FIX (2026-08-09): confirmed the exact, direct root cause of a
+        # live production 500 error on /api/radar/scores -- .sum() on
+        # a pandas boolean Series returns numpy.int64, not a plain
+        # Python int. This value flowed unconverted through min() and
+        # then through the weighted score calculation into the final
+        # dataclass, which FastAPI's jsonable_encoder cannot serialize
+        # (numpy types aren't natively JSON-serializable), breaking
+        # the entire endpoint -- not just this new feature.
+        return int(min(100, failures * 20))
 
     def campaign_continuity_score(self,
                                   df: pd.DataFrame,
